Refactor file removal logic in system-cleanup.py

+++ system-cleanup.py
@@ -30,31 +30,34 @@ try:
                     sizes.append(size)
 
             if len(match) > 0:
+                print(f"Files: {len(files)}")
                 print(f"Match: {len(match)}")
                 match_reference = dict(zip(match, sizes))
                 for k, v in match_reference.items():
                     print(f"{v} {unit} | {k}")
-                remove = input("Remove all files or selection (a/s): ")
-                if remove == 'a':
-                    for item in match:
-                        item.unlink()
-                        print(f"{item} removed successfully")
-                elif remove == 's':
-                    total = int(input("How many files to remove? "))
-                    if 0 < total <= len(match):
-                        nums = [i for i in range(len(match))]
-                        reference = dict(zip(nums, match))
-                        for index, file in reference.items():
-                            print(index, file)
-                        count = 0
-                        while count < total:
-                            select = int(input("Select file (number): "))
-                            if select in reference.keys():
-                                reference[select].unlink()
-                                print(f"{reference[select]} removed successfully")
-                                count += 1
-                            else:
-                                print("Number not in list")
+                decide = input("Remove files? (y/n): ")
+                if decide == 'y':
+                    remove = input("All files or a selection? (a/s): ")
+                    if remove == 'a':
+                        for item in match:
+                            item.unlink()
+                            print(f"{item} removed successfully")
+                    elif remove == 's':
+                        total = int(input("How many files to remove? "))
+                        if 0 < total <= len(match):
+                            nums = [i for i in range(len(match))]
+                            reference = dict(zip(nums, match))
+                            for index, file in reference.items():
+                                print(index, file)
+                            count = 0
+                            while count < total:
+                                select = int(input("Select file (number): "))
+                                if select in reference.keys():
+                                    reference[select].unlink()
+                                    print(f"{reference[select]} removed successfully")
+                                    count += 1
+                                else:
+                                    print("Number not in list")
             else:
                 print(f"Files: {len(files)}")
                 print(f"Match: {len(match)}")
